+++ forui/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.13.0 (Next)
 
 ### Others
+* Fix `FSelect.search(...)` always focusing on 1st item even when there is a selected item.
+* Fix `FSelect.search(...)` expanding items unnecessarily.
 * Fix `FTappable` persisting pressed effect even after pointer is moved outside the widget.